Refactoring pour ne plus utiliser les decorateurs et mise en place des parametre d'url comme etant un path

+++ domain/repo/routes.py
@@ -3,7 +3,6 @@ from flask import request
 from kink import di
 from domain.access.access_interface import IAccess
 from domain.repo.repo_interface import IRepo
-from domain.access.decorators.access_decorator import verifyAccessToken, verifyCanCreate, verifyCreator
 from domain.repo.schemas.repo_schema import RepoListSchema, RepoInputSchema
 
 repo_app = APIBlueprint('repo_app', __name__)
@@ -11,11 +10,15 @@ repo_app = APIBlueprint('repo_app', __name__)
 
 @repo_app.get("/api/repo/list")
 @repo_app.output(RepoListSchema)
-@verifyAccessToken
 def listRepoContributedByUser():
     repo = di[IRepo]
     access = di[IAccess]
-    username = access.decodeAccessToken(request.headers.get("Access-token"))["username"]
+
+    # Verification
+    access_token = request.headers.get("Access-token")
+    access.verifyAccessToken(access_token)
+
+    username = access.decodeAccessToken(access_token)["username"]
     the_repos = repo.getRepoContributedBy(username)
     out_repos = []
     for r in the_repos:
@@ -25,24 +28,31 @@ def listRepoContributedByUser():
 
 @repo_app.post("/api/repo/create")
 @repo_app.input(RepoInputSchema)
-@verifyCanCreate
-@verifyAccessToken
 def createRepo(json_data):
     repo = di[IRepo]
     access = di[IAccess]
-    username = access.decodeAccessToken(request.headers.get("Access-token"))["username"]
+
+    # Verification
+    access_token = request.headers.get("Access-token")
+    access.verifyAccessToken(access_token)
+    access.verifyCanCreate(access_token)
+
+    username = access.decodeAccessToken(access_token)["username"]
     repo.addRepo(json_data["repo_path"], username)
     return {"message": "Repo created"}
 
 
 @repo_app.delete("/api/repo/delete")
 @repo_app.input(RepoInputSchema)
-@verifyAccessToken
 def deleteRepo(json_data):
     access = di[IAccess]
     repo = di[IRepo]
-    repo.verifyRepoExist(json_data['repo_path'])
-    access.verifyCreator(request.headers.get("Access-token"), json_data["repo_path"])
+
+    # Verification
+    access_token = request.headers.get("Access-token")
+    access.verifyAccessToken(access_token)
+    access.verifyCreator(access_token, json_data["repo_path"])
+
     repo.removeRepo(json_data["repo_path"])
     return {"message": "Repo removed"}
